// FrontEnd/ChamberAppel/src/app/generatedapis/models/DtoFiltreChamberAppeleDatatableRequest.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { DtoFiltreChamberAppele } from './DtoFiltreChamberAppele';
import type { DtoPagination } from './DtoPagination';
export type DtoFiltreChamberAppeleDatatableRequest = {
    filtre?: DtoFiltreChamberAppele;
    pagination?: DtoPagination;
};