// FrontEnd/ChamberAppel/src/app/generatedapis/models/ChamberAppele.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { CRC } from './CRC';
export type ChamberAppele = {
    id?: string;
    numero_Dossier?: string | null;
    appelant?: string | null;
    numero_Jugement_Faisant_Objet_De_Appel?: string | null;
    emis_En_Date_Du?: string | null;
    idCRC?: string;
    centre_Comptable?: string | null;
    exercice_fiscal?: string;
    date_Enregistrement_Requete_Cour_Regionale_Des_Comptes?: string | null;
    conseiller_Rapporteur?: string | null;
    date_Ordonnance_Designation_Conseiller_Rapporteur?: string | null;
    conseiller_Rapporteur_Remplacant?: string | null;
    date_Ordonnance_Designation_Conseiller_Rapporteur_Remplacant?: string | null;
    date_Demande_Documents_Supplementaires?: string | null;
    date_Envoi_Requete_Parties?: string | null;
    parties_Agent_du_Roi_au_CRC?: string | null;
    parties_Entrepot_Regional?: string | null;
    parties_Travailleur_du_territoire?: string | null;
    parties_President_de_la_Commune?: string | null;
    date_de_reception_par_les_parties_Agent_du_Roi_au_CRC?: string | null;
    date_de_reception_par_les_parties_Entrepot_Regional?: string | null;
    date_de_reception_par_les_parties_Travailleur_du_territoire?: string | null;
    date_de_reception_par_les_parties_President_de_la_Commune?: string | null;
    reponse_au_memoire_d_appel_Agent_du_Roi_au_CRC?: string | null;
    reponse_au_memoire_d_appel_Entrepot_Regional?: string | null;
    reponse_au_memoire_d_appel_Travailleur_du_territoire?: string | null;
    reponse_au_memoire_d_appel_President_de_la_Commune?: string | null;
    date_Preparation_Rapport?: string | null;
    designation_du_Conseiller_Reviseur?: string | null;
    date_Designation_du_Conseiller_Reviseur?: string | null;
    designation_du_Conseiller_Reviseur_Remplacant?: string | null;
    date_Designation_du_Conseiller_Reviseur_Remplacant?: string | null;
    date_de_preparation_de_l_avis_de_l_examinateur?: string | null;
    date_Transmission_Dossier_Ministere_Public?: string | null;
    date_Retour_Dossier_Ministere_Public?: string | null;
    date_Conclusions_Ministere_Public?: string | null;
    numero_Conclusions_Ministere_Public?: string | null;
    date_de_la_session?: string | null;
    dispositif_Decision?: string | null;
    numero_Decision?: string | null;
    date_d_envoi_de_la_decision_pour_notifier_les_parties?: string | null;
    date_de_reception_par_le_comptable_d_une_copie_de_la_decision?: string | null;
    updatedBy?: string | null;
    updateTime?: string | null;
    isActive?: number;
    crc?: CRC;
};