// FrontEnd/ChamberAppel/src/app/generatedapis/models/ChamberAppeleDatatableResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { ChamberAppele } from './ChamberAppele';
export type ChamberAppeleDatatableResponse = {
    data?: Array<ChamberAppele> | null;
    recordTotal?: number;
    recordFiltred?: number;
};